+++ CHANGELOG.md
@@ -1096,6 +1096,7 @@ activity and on its child transactions (which can be actuals, refunds, and adjus
 
 - Refactor the importers (including associated uploads controllers and views) to use a unified approach for all entities being imported
 - Ensure `Forecast.set_value` always returns a forecast or nil; this will ensure that uploaded forecasts are correctly displayed back to the user on the success page
+- Add Level B budget bulk upload functionality - form with errors/confirmation view; link in top nav
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-117...HEAD
 [release-117]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-116...release-117